updated views to use @api view instead of @login required

+++ accounts/views.py
@@ -5,37 +5,48 @@ from django.http import JsonResponse, HttpResponse
 from django.contrib.auth import get_user_model
 from django.shortcuts import get_object_or_404
 from rest_framework import generics
-from .models import User, Friend_Request
+from .models import User, FriendRequest
 from rest_framework.permissions import IsAdminUser, IsAuthenticatedOrReadOnly
-# Create your views here.
+from rest_framework.decorators import api_view, permission_classes
+from rest_framework.response import Response
+from rest_framework import permissions, status
+
 
 class UserListAPIView(generics.ListCreateAPIView):
-    queryset = User.objects.all();
+    queryset = User.objects.all()
     serializer_class = UserSerializer
 
-@login_required 
-def send_match_request(request, userID):
-    from_user = request.user
-    to_user = User.objects.get(id=userID)
-    friend_request, created = Friend_Request.objects.get_or_create(
-        from_user=from_user, to_user=to_user)
-    if created:
-        return HttpResponse('friend request sent')
-    else:
-        return HttpResponse('friend request already sent')
 
-@login_required
+@api_view(['POST'])
+@permission_classes((permissions.AllowAny,))
+def send_match_request(request, userID):
+    if request.method == 'POST':
+        from_user = request.user
+        to_user = User.objects.get(id=userID)
+        friend_request, created = FriendRequest.objects.get_or_create(
+            from_user=from_user, to_user=to_user)
+        if created:
+            return HttpResponse('friend request sent')
+        else:
+            return HttpResponse('friend request already sent')
+
+@api_view(['POST'])
+@permission_classes((permissions.AllowAny,))
 def accept_match_request(request, requestID):
-    friend_request = Friend_Request.object.get(id=requestID)
-    if friend_request.to_user == request.user:
-        friend_request.to_user.buddies.add(friend_request.from_user)
-        friend_request.from_user.buddies.add(friend_request.to_user)
-        friend_request.delete()
-        return HttpResponse('friend request accepted')
-    else:
-        return HttpResponse('friend request not accepted')
-    
-@login_required
+    if request.method == 'POST':
+        friend_request = FriendRequest.objects.get(id=requestID)
+        if friend_request.to_user == request.user:
+            friend_request.to_user.buddies.add(friend_request.from_user)
+            friend_request.from_user.buddies.add(friend_request.to_user)
+            friend_request.delete()
+            return Response('request accepted')
+        else:
+            return Response('request not accepted')
+
+
+@api_view(['GET'])
+@permission_classes((permissions.AllowAny,))
 def match_request_count(request):
-    count = Friend_Request.objects.filter(to_user=request.user).count()
-    return JsonResponse({'count': count})
+    if request.method == 'GET':
+        count = FriendRequest.objects.filter(to_user=request.user).count()
+    return Response(count, status=status.HTTP_200_OK)
